fix(notification-indexer): robust vote-poll cursor (review findings)
Address two Qodo review findings on the vote-threshold poller.

- Non-monotonic cursor (correctness): votes_count.updated_at was set with now()
  (transaction-start time), so under concurrent vote-indexer writers a row could
  commit with an updated_at behind the poller's keyset cursor and be missed
  forever. The vote-indexer now stamps clock_timestamp() (statement wall-clock),
  and the poller re-scans a bounded trailing overlap window each tick
  (VOTE_POLL_OVERLAP_SECS, default 10s) with a high-water guard that never moves
  the cursor backward. Re-scanned rows are deduped by the already_notified
  anti-check + idempotency key, so the overlap is free.
- Keyset type mismatch (perf): the cursor binds i64 but votes_count.id is int4;
  cast the keyset bound to ::int so the (updated_at, id) index is used as-is.

README documents VOTE_POLL_OVERLAP_SECS. 82 unit + 29 vote-indexer + 198 e2e green.

+++ notification-service/notification-indexer/src/storage.rs
@@ -618,7 +618,9 @@ impl Storage {
                    ) AS already_notified
             FROM votes_count vc
             WHERE vc.object_type = 0
-              AND (vc.updated_at, vc.id) > ($1, $2)
+              -- $2::int matches vc.id (int4) so the (updated_at, id) keyset index is
+              -- used as-is; cursor_id is always within int4 range (from votes_count.id).
+              AND (vc.updated_at, vc.id) > ($1, $2::int)
             ORDER BY vc.updated_at, vc.id
             LIMIT $4
             "#,
